// src/types/dragItemViewType.ts
/* eslint-disable no-unused-vars */
export enum DragItemViewType {
  DcodeBanner = 'DcodeBanner',
  DcodeLocation = 'DcodeLocation',
  DcodeGrid = 'DcodeGrid',
  DcodeMember = 'DcodeMember',
  DcodeAdvertising = 'DcodeAdvertising',
  DcodeSearchBar = 'DcodeSearchBar',
  DcodeProdRecomPro = 'DcodeProdRecomPro',
  DcodeSecondsKill = 'DcodeSecondsKill',
  DcodeSpellGroup = 'DcodeSpellGroup',
  DcodeCoupons = 'DcodeCoupons',
  DcodeOrderActive = 'DcodeOrderActive',
  DcodeDiscountTime = 'DcodeDiscountTime',
  DcodePointsMerchandise = 'DcodePointsMerchandise',
  DcodeExchange = 'DcodeExchange',
  DcodeLevitationTool = 'DcodeLevitationTool',
  DcodeDesignCustorm = 'DcodeDesignCustorm',
  DcodePaddingLine = 'DcodePaddingLine',
  DcodeHotZone = 'DcodeHotZone',
  DcodeLiveStreaming = 'DcodeLiveStreaming',
  DcodeNoticeBar = 'DcodeNoticeBar',
  DcodeShareBtn = 'DcodeShareBtn',
  DcodeWeCom = 'DcodeWeCom',
}

export const DropAcceptList = [
  DragItemViewType.DcodeBanner,
  DragItemViewType.DcodeLocation,
  DragItemViewType.DcodeGrid,
  DragItemViewType.DcodeMember,
  DragItemViewType.DcodeAdvertising,
  DragItemViewType.DcodeSearchBar,
  DragItemViewType.DcodeProdRecomPro,
  DragItemViewType.DcodeSecondsKill,
  DragItemViewType.DcodeSpellGroup,
  DragItemViewType.DcodeCoupons,
  DragItemViewType.DcodeOrderActive,

  DragItemViewType.DcodeDiscountTime,
  DragItemViewType.DcodePointsMerchandise,
  DragItemViewType.DcodeExchange,
  DragItemViewType.DcodeLevitationTool,
  DragItemViewType.DcodeDesignCustorm,

  DragItemViewType.DcodePaddingLine,
  DragItemViewType.DcodeHotZone,
  DragItemViewType.DcodeLiveStreaming,
  DragItemViewType.DcodeNoticeBar,
  DragItemViewType.DcodeShareBtn,
  DragItemViewType.DcodeWeCom,
]

export enum DragType {
  Add = 'Add',
  Adjust = 'Adjust',
}